Allow the setting of content type priority
The negotiation priorities could be different for different cases.
If a user agent sends a request with the following header:

Accept: application/json, text/plain, */*

The response will be HTML, because `*/*` matches `text/html` which is top
priority by default in the format negotiator.

The user agent expects a json response over a html
response though. Maybe the negotiation in general should change, but
being able to set the priority helps to circumvent the issue.